requirements.txt
pika==0.11.0
robotframework==3.0.2
paramiko==2.1.1
pycrypto==2.6.1
robotframework-sshlibrary==2.1.3
robotframework-scplibrary==1.1.1